Restructure scheduler service
Task and Scheduler have been split into different classes and renamed to
better suit the use cases.
