+++ src/main/kotlin/dev/mizarc/waystonewarps/WaystoneWarps.kt
@@ -12,6 +12,7 @@ import dev.mizarc.waystonewarps.application.actions.warp.RefreshAllStructures
 import dev.mizarc.waystonewarps.application.actions.warp.UpdateWarpIcon
 import dev.mizarc.waystonewarps.application.actions.warp.UpdateWarpName
 import dev.mizarc.waystonewarps.application.services.*
+import dev.mizarc.waystonewarps.application.services.scheduling.SchedulerService
 import dev.mizarc.waystonewarps.domain.discoveries.DiscoveryRepository
 import dev.mizarc.waystonewarps.domain.playerstate.PlayerStateRepository
 import dev.mizarc.waystonewarps.domain.warps.WarpRepository
@@ -24,11 +25,11 @@ import dev.mizarc.waystonewarps.infrastructure.persistence.discoveries.Discovery
 import dev.mizarc.waystonewarps.infrastructure.persistence.playerstate.PlayerStateRepositoryMemory
 import dev.mizarc.waystonewarps.infrastructure.persistence.storage.SQLiteStorage
 import dev.mizarc.waystonewarps.infrastructure.persistence.warps.WarpRepositorySQLite
-import dev.mizarc.waystonewarps.infrastructure.services.BukkitSchedulerService
 import dev.mizarc.waystonewarps.infrastructure.services.MovementMonitorServiceBukkit
 import dev.mizarc.waystonewarps.infrastructure.services.PlayerAttributeServiceVault
 import dev.mizarc.waystonewarps.infrastructure.services.StructureBuilderServiceBukkit
 import dev.mizarc.waystonewarps.infrastructure.services.TeleportationServiceBukkit
+import dev.mizarc.waystonewarps.infrastructure.services.scheduling.SchedulerServiceBukkit
 import dev.mizarc.waystonewarps.interaction.listeners.*
 import org.koin.core.context.startKoin
 import org.koin.dsl.module
@@ -50,8 +51,8 @@ class WaystoneWarps: JavaPlugin() {
     private lateinit var playerAttributeService: PlayerAttributeService
     private lateinit var structureBuilderService: StructureBuilderService
     private lateinit var teleportationService: TeleportationService
-    private lateinit var configService: ConfigServiceBukkit
-    private lateinit var scheduler: Scheduler
+    private lateinit var configService: ConfigService
+    private lateinit var scheduler: SchedulerService
 
     override fun onEnable() {
         logger.info(Chat::class.java.toString())
@@ -83,7 +84,7 @@ class WaystoneWarps: JavaPlugin() {
         configService = ConfigServiceBukkit(this.config)
         playerAttributeService = PlayerAttributeServiceVault(configService, metadata)
         structureBuilderService = StructureBuilderServiceBukkit()
-        scheduler = BukkitSchedulerService(this)
+        scheduler = SchedulerServiceBukkit(this)
         teleportationService = TeleportationServiceBukkit(playerAttributeService, movementMonitorService, scheduler)
     }
