+++ src/main/kotlin/dev/mizarc/waystonewarps/application/services/scheduling/SchedulerService.kt
@@ -1,0 +1,8 @@
+package dev.mizarc.waystonewarps.application.services.scheduling
+
+/**
+ * Schedules an event to run after an X amount of time
+ */
+interface SchedulerService {
+    fun schedule(delayTicks: Long, task: () -> Unit): Task
+}
