+++ src/main/kotlin/dev/mizarc/waystonewarps/application/services/scheduling/Task.kt
@@ -1,0 +1,8 @@
+package dev.mizarc.waystonewarps.application.services.scheduling
+
+/**
+ * Represents a task that is to be performed
+ */
+interface Task {
+    fun cancel()
+}
