+++ src/main/kotlin/dev/mizarc/waystonewarps/infrastructure/services/scheduling/TaskBukkit.kt
@@ -1,0 +1,10 @@
+package dev.mizarc.waystonewarps.infrastructure.services.scheduling
+
+import dev.mizarc.waystonewarps.application.services.scheduling.Task
+import org.bukkit.scheduler.BukkitRunnable
+
+class TaskBukkit(private val runnable: BukkitRunnable) : Task {
+    override fun cancel() {
+        runnable.cancel()
+    }
+}
